search: handle search_complete() properly
We create a tab for the search results when we submit the search.
Upon completion, the tab will be filled with (track) results of
said search.

Callbacks will be set, so playback of the results will also work.

Note that we don't do search continuation right now, so
SPTHUI_SEARCH_CHUNK results is all you get right now..