ui: skeleton for item popup menus
The idea is that we offer to go to the specified track / album /
artist / whatever when the user right-clicks an item.

Shove the code to do this into a separate helper module. Things
will still need restructuring if we want to be able to do
tab_add() and things from the callbacks..