spthui: try to obey libspotify locking rules
Our lock ordering for now is

  1. Global gdk lock (gdk_threads_enter())
  2. spthui->lock (guards sp_*())

We start holding spthui->lock around sp_process_events() to
obey this.

This means that the callbacks called via sp_process_events()
need to be aware of the lock and drop it in case they need to
gdk_threads_{enter,leave}().

GTK signal handlers are called with the global gdk lock held, so
if they end up calling and sp_*() functions, they need to grab
spthui->lock. But they don't need to drop the gdk lock.