spthui: fix search setup + callback arguments
Search has been broken since we've been doing sp_*_add_ref()s
in item_init_*() constructors.

This was due to a thinko where we passed a struct item * embedding
the search object to the search callback. Problem was, we started
pinning sp_* objects in the various item_init_*() constructors.

We had a struct search helper for wrapping sp_search, and we'd
embed this in a struct item. Problem was we needed the item
for constructing the sp_search itself, so obviously item_init_search()
couldn't have the sp_search available and cheerfully segfaulted.

Fix it by passing the struct search as data for the ->search_complete()
callback.