playback_panel: show position tooltip when hovering over progress bar
Show a position tooltip for estimating seek position when the user
is hovering a pointer over the progress bar.

GTK memory management is confusing as always, but we seem to get away
with building the tip on stack and gtk_tooltip_set_text()ing it in the
"query-tooltip" signal handler.

So I suppose it makes a copy then.