spthui: construct search target view early
Creating the search target view after the search itself isn't
a terribly bright idea. We end up racing the search completion
with view creation and the search tends to win it for any rapid
clickery.

So make sure the view (and its store) is in place before we hit
the search. The tab itself doesn't really matter, it will appear
eventually.