popup: fix label reallocation for long labels
We can't just realloc() + re-vsnprintf() with the same va_list
in case the title ends up being too long on the first go.

Fix this by indicating to the caller whether we were able to
set the label up or not, so the caller can reset va_list and
redo.

It results in a rather horrible calling convention, though...