item: do not copy name in item_init_*(..., name)
Change the item_init_(..., name) calling conventions to not
copy name, but store a reference instead.

item_free() will release the name, however, so the caller should
generally pass in a freshly allocated string.

A lot of callers tend to construct item names dynamically on demand
and copying it leads to endless tiny allocations and back-and-forth
copies.