playback_panel: reset ->duration on playback start unconditionally
Fix a bug where we'd set panel->duration to 0 when playback was
stopped with a current track gone missing (playback_panel_set_info()
with track=NULL), but wouldn't reset it to track duration if the _same_
track was restarted.

The whole track-changed playback-stopped-started thing would need
some sort of an overhaul, it's starting to get not funny.