search: get rid of the local ITER_* enumeration, use ITEM_* instead
The search-local ITER_* enumeration for identifying result item
types is redundant, as we already have ITEM_* available. The space
wasted for items not relevant to search_*() is negligible.

To help array sizing introduce a new ITEM__COUNT constant that
sits at the end of the item enumeration. This results in useless
case statements wherever we switch() over item_type(), but I
suppose we can live with that.