spthui: scale the album image to an arbitary size for tab headers
Use 15 as a handy-wavy height for scaling album art into tab headers.

In a somewhat weird turn of events we apparently should no longer
g_object_unref() the loader after we unref the only image loaded
after with it (the original, once we have the scaled version and no
longer need the unscaled one).

This must mean it's somehow pinning the loader and when we unref it
it lets go.