spthui: propagate errors from track_play()
We don't indicate unplayable tracks on an album in any manner.
Loading such a track will, however, fail.

Propagate these errors from track_play() so the ->end_of_track()
navigation can skip these and move on to the next one.

Yeah, we'll end up looping the list forever in case something
catastrophic happens and nothing on the album/playlist is playable.
That'll get sorted out once it becomes a problem.