playback_panel: reset position on current track restart
If the currently playing track is restarted, we want to reset
the progress bar.

The old heuristics won't catch it as they only see playback starting
on the same track, which is indistinguishable from pause.

The caller knows how to deal with this, so provide an api for
resetting progress.