mpris: don't pass messages to other handlers if we already handled it
If we can eval() a symbol, we can consider it handled and not pass it
on to other handlers.

Additionally, don't handle mpris_symbol_eval() errors in mpris.c anymore.
By the time we have a symbol, it should be able to report evaluation
errors all by itself, possibly sending dbus error messages as a result.