mpris: beginnings of org.freedesktop.mpris.MediaPlayer2.Player
Implement the first effective dbus/mpris callback and behold,
we now support org.freedesktop.mpris.MediaPlayer2.Player.PlayPause.

Where there's a simple toggle UI element it seems to work nicely,
as we can just gtk_widget_activate() it and the UI reflects the state
change nicely. The only semi-ugly part of it is the bouncing in the
callback in order to get something suitable for g_idle_add(), but
that's livable with and can get cleaned up later.

And there's something to fix in our spthui core too, as playback
toggling doesn't seem to do anything if there's nothing selected
to begin with. I suppose for remote control we'd like to start
playback when there's a current view, even though nothing on it
is selected.

There's also some racing going on as we export the DBus interfaces
very early and don't have the UI set up and shown at that time. If
there's a DBus call while we're setting up the UI, things will
probably go terribly wrong.