mpris/properties: DBusMessageIter-based model for *_get()
Swap the DBusMessage * argument in *_get() for an DBusMessageIter *
one.

This should make implementing *.DBus.Properties.GetAll() a bit easier
as the common code can do the "array of {sv}" part and loop over the
handlers passing a proper container iterator to *_get().

The *_get() implementations themselves won't need to know if they're
called specifically or if someone is just iterating through everything.