gtk: compat_* wrapping for various STOCK_* bits deprecated in 3.10
GTK 3.10 deprecated a bunch of the "stock item" parts we rely on for
button icons and the like.

Since we're rather strict about it and use CFLAGS to disable everything
that's not in vogue in the GTK land any given day, our builds fail when
we hit this deprecated functionality.

Fortunately it's trivial to emulate / provide wrappings for, but I have
to say compat_gtk.h is getting hilarious.