ui: implement compat wrappers to deal with deprecated GtkAlignment
Gtk 3.14 deprecated GtkAlignment as means of... doing alignment,
and promotes using widget halign, valign and margin properties
on the widget to achieve the same effect.

Fine.

Implement the alignment bits we have as compat_gtk_*() wrappers.
GTK2 gets the current GtkAlignment code, GTK3 gets the properties
approach. The necessary properties are documented "since 3.0" so
no need to check for 3.14 specifically.

Of course the new alignment things are a bit.. coarser there, but
given how the UI looks anyway, nobody's going to care much.

We now also get real compat_$(UI).o objects for the compat gunk, since
doing these as inlines in the header is a bit much. Plus it leads to
warnings if they're not used everywhere they're included.