compat: add compat_gtk_try_force_button_images() for.. forcing images
GTK 3.6 deprecated the gtk-button-images GtkSetting, in favour of the
per-button always-show-image property.

It seems on the GTK 3 version we're testing (2.26) button images are
shown by default, and always-show-image isn't needed either.

But if it turns out it will be needed, we'll just provide new wrappers
on top of our COMPAT_GTK_*_BUTTON() media button... wrappers.